change double type to BigDecimal on account balance

+++ transaction-service/src/main/java/ru/banking/dto/AccountDtoResponse.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,5 +13,5 @@ public class AccountDtoResponse {
 
     Long id;
 
-    Double balance;
+    BigDecimal balance;
 }
